Update painting with AI
Odds and probability One of the reasons why modern AI methods actually work in real-world problems is their ability to deal with uncertainty. probability can also be used to quantify and compare risks in everyday life by systematically collecting data, we can critically evaluate probabilistic statements, and our numbers can sometimes be found to be right or wrong. In other words, the key lesson is that uncertainty is not beyond the scope of rational thinking and discussion, and probability provides a systematic way of doing just that. Odds are a numerical expression, typically represented by a pair of numbers, used both in gambling and statistics as a way of understanding probability. Machine learning applies the principle of odds and probability in making decision odds help inform the predictive accuracy of neutral networks s e.g. improving things like sports predictions to better help club owners and managers price their bets.